third_party/WebKit/LayoutTests/media/video-controls-captions-on-off.html

<!DOCTYPE html>
<title>Tests that tracks can be turned on and off through the track selection menu.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<script src="media-file.js"></script>
<video controls></video>
<script>
async_test(function(t) {
    var captions = ["First", "Second", "Third"];
    var video = document.querySelector("video");

    video.oncanplaythrough = t.step_func_done(function() {
        var track1 = video.addTextTrack("captions");
        var track2 = video.addTextTrack("captions");

        for (var i = 0; i < captions.length; i++) {
            track1.addCue(new VTTCue(0, 120, captions[i]));
            track2.addCue(new VTTCue(0, 120, captions[i]));
        }
        assert_true(isClosedCaptionsButtonVisible(video));

        // The captions track should be listed in textTracks, but not yet loaded.
        assert_equals(video.textTracks.length, 2);
        assert_equals(video.textTracks[0].mode, "hidden");
        checkCaptionsHidden(video);

        // Captions track should become visible after the track is selected.
        clickTextTrackAtIndex(video, 0);
        checkCaptionsVisible(video, captions);

        // Captions should not be visible after they're turned off through the menu.
        turnClosedCaptionsOff(video);
        checkCaptionsHidden(video);

        // Captions track should become visible after the track is selected again.
        clickTextTrackAtIndex(video, 0);
        checkCaptionsVisible(video, captions);
    });

    video.src = findMediaFile("video", "content/counting");
});
</script>
